+++ fabric/src/main/java/de/maxhenkel/voicechat/FabricVoicechatClientMod.java
@@ -3,7 +3,6 @@ package de.maxhenkel.voicechat;
 import com.sun.jna.Platform;
 import de.maxhenkel.configbuilder.ConfigBuilder;
 import de.maxhenkel.voicechat.config.FabricClientConfig;
-import de.maxhenkel.voicechat.integration.ClothConfig;
 import de.maxhenkel.voicechat.macos.VersionCheck;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
@@ -17,7 +16,6 @@ public class FabricVoicechatClientMod extends VoicechatClient implements ClientM
     public void onInitializeClient() {
         CLIENT_CONFIG = ConfigBuilder.build(Minecraft.getInstance().gameDirectory.toPath().resolve("config").resolve(Voicechat.MODID).resolve("voicechat-client.properties"), true, FabricClientConfig::new);
         initializeClient();
-        ClothConfig.init();
 
         if (Platform.isMac() && !CLIENT_CONFIG.javaMicrophoneImplementation.get()) {
             CLIENT_CONFIG.javaMicrophoneImplementation.set(true).save();