+++ forge/src/main/java/de/maxhenkel/voicechat/intercompatibility/ForgeClientCompatibilityManager.java
@@ -3,6 +3,8 @@ package de.maxhenkel.voicechat.intercompatibility;
 import com.mojang.blaze3d.platform.InputConstants;
 import de.maxhenkel.voicechat.events.ClientVoiceChatConnectedEvent;
 import de.maxhenkel.voicechat.events.ClientVoiceChatDisconnectedEvent;
+import de.maxhenkel.voicechat.integration.clothconfig.ClothConfigIntegration;
+import de.maxhenkel.voicechat.integration.clothconfig.ForgeClothConfigIntegration;
 import de.maxhenkel.voicechat.voice.client.ClientVoicechatConnection;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
@@ -33,6 +35,7 @@ public class ForgeClientCompatibilityManager extends ClientCompatibilityManager 
     private final List<RenderHUDEvent> renderHUDEvents;
     private final List<KeyboardEvent> keyboardEvents;
     private final List<MouseEvent> mouseEvents;
+    private final List<Runnable> clientTickEvents;
     private final List<Runnable> inputEvents;
     private final List<Runnable> disconnectEvents;
     private final List<Runnable> joinServerEvents;
@@ -47,6 +50,7 @@ public class ForgeClientCompatibilityManager extends ClientCompatibilityManager 
         renderHUDEvents = new ArrayList<>();
         keyboardEvents = new ArrayList<>();
         mouseEvents = new ArrayList<>();
+        clientTickEvents = new ArrayList<>();
         inputEvents = new ArrayList<>();
         disconnectEvents = new ArrayList<>();
         joinServerEvents = new ArrayList<>();
@@ -78,6 +82,14 @@ public class ForgeClientCompatibilityManager extends ClientCompatibilityManager 
     @SubscribeEvent
     public void onMouse(InputEvent.RawMouseEvent event) {
         mouseEvents.forEach(mouseEvent -> mouseEvent.onMouseEvent(minecraft.getWindow().getWindow(), event.getButton(), event.getAction(), event.getMods()));
+    }
+
+    @SubscribeEvent
+    public void onKeyInput(TickEvent.ClientTickEvent event) {
+        if (event.phase != TickEvent.Phase.START) {
+            return;
+        }
+        clientTickEvents.forEach(Runnable::run);
     }
 
     @SubscribeEvent
@@ -152,6 +164,11 @@ public class ForgeClientCompatibilityManager extends ClientCompatibilityManager 
     }
 
     @Override
+    public void onClientTick(Runnable onClientTick) {
+        clientTickEvents.add(onClientTick);
+    }
+
+    @Override
     public InputConstants.Key getBoundKeyOf(KeyMapping keyBinding) {
         return keyBinding.getKey();
     }
@@ -217,5 +234,10 @@ public class ForgeClientCompatibilityManager extends ClientCompatibilityManager 
     @Override
     public void addResourcePackSource(PackRepository packRepository, RepositorySource repositorySource) {
         packRepository.addPackFinder(repositorySource);
+    }
+
+    @Override
+    public ClothConfigIntegration getClothConfigIntegration() {
+        return new ForgeClothConfigIntegration();
     }
 }